TLRC-284 Create 'Your T Levels' page  Code reivew comments are incorporated.

+++ src/Sfa.Tl.ResultsAndCertification.Api/Controllers/AwardingOrganisationController.cs
@@ -26,8 +26,7 @@ namespace Sfa.Tl.ResultsAndCertification.InternalApi.Controllers
         public async Task<IEnumerable<AwardingOrganisationPathwayStatus>> GetAllTlevelsByAwardingOrganisationIdAsync()
         {
             // TODO: following statement to be updated?
-            var id = !string.IsNullOrEmpty(User.GetUkPrn()) ? int.Parse(User.GetUkPrn()) : 10009696;
-           
+            var id = !string.IsNullOrEmpty(User.GetUkPrn()) ? long.Parse(User.GetUkPrn()) : 10009696;
             return await _awardingOrganisationService.GetAllTlevelsByAwardingOrganisationIdAsync(id);
         }
     }
